ibs: move to klplib package
