+++ klpbuild/main.py
@@ -5,14 +5,13 @@
 
 import sys
 
-from klpbuild.klplib.codestream import Codestream
 from klpbuild.extractor import Extractor
-from klpbuild.ibs import IBS
 from klpbuild.inline import Inliner
+from klpbuild.klplib.cmd import create_parser
+from klpbuild.klplib.codestream import Codestream
+from klpbuild.klplib.ibs import IBS
 from klpbuild.klplib.ksrc import GitHelper
 from klpbuild.plugins.setup import Setup
-from klpbuild.klplib.cmd import create_parser
-from klpbuild.klplib.ksrc import GitHelper
 
 
 def main():
